goierjgiodfsjgdfsgdfsgdfijgosdfg

+++ database/migrations/2023_03_16_002431_add_ffunccod_to_mfos.php
@@ -15,7 +15,7 @@ class AddFfunccodToMfos extends Migration
     {
         Schema::table('major_final_outputs', function (Blueprint $table) {
             //
-            $table->bigInteger('FFUNCCOD')->nullable();
+            $table->Integer('FFUNCCOD')->nullable();
         });
     }
